+++ airborn.web/Models/Aircraft/C172_SP.cs
@@ -9,18 +9,18 @@ namespace Airborn.web.Models
         {
         }
 
-        public override double MakeTakeoffAdjustments(PerformanceCalculationResult result, double takeoffDistance)
+        public override decimal MakeTakeoffAdjustments(PerformanceCalculationResult result, decimal takeoffDistance)
         {
 
             if (result.HeadwindComponent > 0)
             {
                 // subtract 10% for each 9 knots of headwind
-                takeoffDistance = takeoffDistance * (1 - ((result.HeadwindComponent / 9) * 0.1));
+                takeoffDistance = takeoffDistance * (1 - ((result.HeadwindComponent / 9) * 0.1m));
             }
             else if (result.HeadwindComponent < 0)
             {
                 // add 10% for each 2 knots of tailwind up to 10 knots
-                double adjustment = (result.HeadwindComponent / 2) * 0.1;
+                decimal adjustment = (result.HeadwindComponent / 2) * 0.1m;
 
                 takeoffDistance = takeoffDistance * (1 - adjustment);
 
@@ -31,17 +31,17 @@ namespace Airborn.web.Models
             return takeoffDistance;
         }
 
-        public override double MakeLandingAdjustments(PerformanceCalculationResult result, double landingDistance)
+        public override decimal MakeLandingAdjustments(PerformanceCalculationResult result, decimal landingDistance)
         {
             if (result.HeadwindComponent > 0)
             {
                 // subtract 10% for each 9 knots of headwind
-                landingDistance = landingDistance * (1 - ((result.HeadwindComponent / 9) * 0.1));
+                landingDistance = landingDistance * (1 - ((result.HeadwindComponent / 9) * 0.1m));
             }
             else if (result.HeadwindComponent < 0)
             {
                 // add 10% for each 2 knots of tailwind up to 10 knots
-                double adjustment = (result.HeadwindComponent / 2) * 0.1;
+                decimal adjustment = (result.HeadwindComponent / 2) * 0.1m;
 
                 landingDistance = landingDistance * (1 - adjustment);
